@@ -8,14 +8,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-chrono": "^1.21.4",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.8.1",
     "react-toastify": "^8.1.1",
     "sass": "^1.49.7"
   },
   "devDependencies": {
     "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react-dom": "^17.0.19",
     "@vitejs/plugin-react": "^1.0.7",
     "typescript": "^4.5.4",
     "vite": "^2.8.0"
